backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
opencv-python==4.10.0.84
numpy==2.2.2
Pillow==11.1.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
aiofiles==24.1.0
